% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_kitchen.R
\name{boot_kitchen}
\alias{boot_kitchen}
\title{Bootstrap kitchen sink model predictions}
\usage{
boot_kitchen(
  trainx,
  trainy,
  predictx,
  featuresweep,
  windowsweep,
  reps = 100,
  verbose = TRUE
)
}
\arguments{
\item{trainx}{A matrix of independent data to use for training models.
trainx[i,] should correspond with trainy[i].}

\item{trainy}{A vector of dependent data to use for training models.
trainy[i] should correspond with trainx[i,].}

\item{predictx}{A vector or matrix which `boot_kitchen` will predict values for.
Values must correspond with the training data.}

\item{featuresweep}{A value or vector of feature counts to use in the
kitchen sink models.}

\item{windowsweep}{A value or vector of window sizes to use in the
kitchen sink models. Values should not exceed the column length
of the data.}

\item{reps}{The number of repetitions to perform for each combination of
feature count and window size.}

\item{verbose}{Print bootstrap progress.}
}
\value{
Returns a list where index [[f]][[w]] returns the predicted
  values from a model with feature count featuresweep[f] and window size
  windowsweep[w].
  #'
}
\description{
`boot_kitchen` trains models on training data and then predicts values
for some other data using unique normal matrices. The output can be used
for convidence intervals or probability.
}
\details{
This function bootstraps predictions of a kitchen sink model
by fitting models for a number of normal matrices and then predicting
values on the prediction data for each. It is most simple to choose a
single set of superparameters (feature count and window size) ahead of
time, but `boot_kitchen` can perform sweeps across superparameters as well.
}
\examples{
x <- matrix(sample(1:10,10000,T),2000,5)
y <- 5*x[,1] + 20*x[,1]*x[,2] + 3*x[,3]^2 - 10*x[,4] - 2*x[,5]
quicksweep(x[1:1000,],y[1:1000],
   x[1001:2000,],y[1001:2000],
   2^(4:8),2:5)

a <- matrix(sample(1:10,5,T),1,5)
b <- 5*a[,1] + 20*a[,1]*a[,2] + 3*a[,3]^2 - 10*a[,4] - 2*a[,5]
mybootstrap <- boot_kitchen(x,y,a,64,5,100)
hist(unlist(mybootstrap))

}
\seealso{
\code{\link{sweep_kitchen}}()
}
\author{
Avery Kruger
}
