% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamp.R
\name{clamp}
\alias{clamp}
\title{Clamps values beyond a threshold to the threshold.}
\usage{
clamp(x, bias = 0, value = 0, logic = `<`)
}
\arguments{
\item{x}{A numeric object.}

\item{bias}{A hyperparameter that can be used to adjust x.}

\item{value}{The value to set (clamp) other values to.}

\item{logic}{A logical function that returns which in x should be clamped.}
}
\value{
Returns an object formatted identical to x
}
\description{
This function returns its input, with all values beyond a
threshold set to a specified value (default 0).
}
\examples{
x <- matrix(rnorm(16,500,TRUE),4,4)
clamp(x)
clamp(x, logic = `>`)

}
\author{
Avery Kruger
}
