% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitchen.R
\name{kitchen}
\alias{kitchen}
\title{Nonlinearize data with a convolutional kitchen sink}
\usage{
kitchen(data, norm)
}
\arguments{
\item{data}{The data to be nonlinearized. Should not include the
variable to be modeled by the kitchen sink.}

\item{norm}{A normal matrix of dimensions c(window size, feature count).
Can be generated using make_norms(). Window size should be no larger than
the size (column length) of the data.}
}
\value{
Returns a matrix of nonlinearized counterparts to the original rows
  of the provided data. Column length is equal to the number of features,
  as determined by the provided normal matrix.
  #'
}
\description{
This function nonlinearizes data with a (convolutional)
kitchen sink. It matrix multiplies data with a normal matrix
that must be provided, clamping the negative results to 0, and
then averaging across windows. Feature count and window size are
determined by the provided normal matrix.
}
\examples{
x <- matrix(sample(1:10,500,T),100,5)
y <- x[,1]*x[,2]^2-0.5*x[,3]*x[,4]+x[,5]*x[,1]

mynorm <- make_norms(64,5)[[1]][[1]]
nonlinx <- nonlinearize(x, mynorm)
summary(lm(y ~ nonlinx))

}
\seealso{
\code{\link{make_norms}}()
}
\author{
Avery Kruger
}
