% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitchen_sweep.R
\name{kitchen_sweep}
\alias{kitchen_sweep}
\title{Quickly check performance of different kitchen sink models}
\usage{
kitchen_sweep(
  trainx,
  trainy,
  valx,
  valy,
  featuresweep,
  windowsweep,
  clampoutliers = TRUE,
  verbose = FALSE,
  show.plot = FALSE,
  seed = NULL,
  partial_files = NULL,
  ...
)
}
\arguments{
\item{trainx}{A matrix of independent data the kitchen sink models will
be trained on.}

\item{trainy}{A vector of dependent data the kitchen sink models will
be trained on. trainy[i] should correspond to trainx[i,].}

\item{valx}{A matrix of independent data the kitchen sink models will
use to  make predictions.  Columns should be organized identically to the
trainx data.}

\item{valy}{A vector of dependent data the kitchen sink models will use
to validate predictions. valy[i] should correspond to valx[i,]}

\item{featuresweep}{A vector of feature counts to sweep across.}

\item{windowsweep}{A vector of window sizes to sweep across.}

\item{verbose}{Print progress.}

\item{show.plot}{Draw plots of predicted values versus true for each model.}

\item{...}{Arguments to be passed to \code{\link{kitchen_sink}}().}
}
\value{
Returns a matrix of the adjusted R^2 values from the linear models
lm(valy ~ predictions(valx)) for each model. Index [f,w] returns
R^2 of the model for featuresweep[f] and windowsweep[w].
}
\description{
Sweeps across a set of feature counts and window sizes to provide
an idea of how different convolutional kitchen sink models perform.
Trains on a set of training data, x and y, and validates on separate data.
Only uses a single normal matrix for each model, so expect some
variance.
}
\details{
Models are trained by ridge regression using \code{\link[glmnet]{cv.glmnet}},
allowing for reduction of overfitting.
}
\examples{
x <- matrix(sample(1:10,1000,T),200,5)
y <- x[,1]*x[,2]^2-0.5*x[,3]*x[,4]+x[,5]*x[,1]*x[,3]-x[,3]^2*x[,2]

kitchen_sweep(trainx = x[1:100,],
trainy = y[1:100],
valx = x[101:200,],
valy = y[101:200],
featuresweep = c(2^(4:7),2^11,4000,6000,7000,2^13),
windowsweep = 2:5,
verbose = T,
ncores=2)

}
\seealso{
\code{\link{make_norms}}()

  \code{\link{boot_kitchen}}

  \code{\link{kitchen_sink}}()
}
\author{
Avery Kruger
}
