% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_norms.R
\name{make_norms}
\alias{make_norms}
\title{Create a set of normal matrices}
\usage{
make_norms(featuresweep, windowsweep, mean = 0, sd = 1)
}
\arguments{
\item{featuresweep}{A vector of feature sizes. Often varied in
exponents of 2, e.g. 2^c(4:10)}

\item{windowsweep}{A vector of window sizes. Maximum should be no
larger than the length of the data to be analyzed.}

\item{mean}{Mean of the normal distribution.}

\item{sd}{Standard deviation of the normal distribution}
}
\value{
Returns a list of matrices where index [[f]][[w]] returns
the normal matrix for featuresweep[f] and windowsweep[w].
}
\description{
`make_norms` creates a list of normal matrices of mean 0 and
standard deviation 1 for each combination of feature length and
window size that can easily be applied to \code{\link{kitchen_sink}}().
}
\examples{
x <- matrix(1:5000,100,50)

myfeaturesweep <- 2^c(4:10)
mywindowsweep <- seq(10,ncol(x),10)
mynorms <- make_norms(myfeaturesweep, mywindowsweep)

for(f in 1:length(myfeaturesweep)){
  for(w in 1:length(mywindowsweep)){
      kitchen_sink(x,mynorms[[f]][[w]])
      }}

}
\seealso{
\code{\link{kitchen_sweep}}()

\code{\link{kitchen_sink}}()
}
\author{
Avery Kruger
}
