% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep_kitchen.R
\name{sweep_kitchen}
\alias{sweep_kitchen}
\title{Quickly check performance of different kitchen sink models}
\usage{
sweep_kitchen(
  trainx,
  trainy,
  valx,
  valy,
  featuresweep,
  windowsweep,
  verbose = FALSE
)
}
\arguments{
\item{trainx}{A matrix of independent data the kitchen sink models will
be trained on.}

\item{trainy}{A vector of dependent data the kitchen sink models will
be trained on. trainy[i] should correspond to trainx[i,].}

\item{valx}{A matrix of independent data the kitchen sink models will
use to  make predictions.  Columns should be organized identically to the
trainx data.}

\item{valy}{A vector of dependent data the kitchen sink models will use
to validate predictions. valy[i] should correspond to valx[i,]}

\item{featuresweep}{A vector of feature counts to sweep across.}

\item{windowsweep}{A vector of window sizes to sweep across.}

\item{verbose}{Show progress if it takes a while.}
}
\value{
Returns a matrix of the adjusted R^2 values from the linear models
lm(valy ~ predictions(valx)) for each model. Index [f,w] returns
R^2 of the model for featuresweep[f] and windowsweep[w].
}
\description{
Sweeps across a set of feature counts and window sizes to provide
an idea of how different convolutional kitchen sink models perform.
Trains on a set of training data, x and y, and validates on separate data.
Only uses a single normal matrix for each model, so expect some
variance.
}
\examples{
x <- matrix(sample(1:10,500,T),100,5)
y <- x[,1]*x[,2]^2-0.5*x[,3]*x[,4]+x[,5]*x[,1]*x[,3]-x[,3]^2*x[,2]

a <- matrix(sample(1:10,500,T),100,5)
b <- x[,1]*x[,2]^2-0.5*x[,3]*x[,4]+x[,5]*x[,1]*x[,3]-x[,3]^2*x[,2]

quicksweep(x,y,a,b,2^(4:8),2:5)

}
\seealso{
\code{\link{make_norms}}()

  \code{\link{boot_kitchen}}

  \code{\link{kitchen}}()
}
\author{
Avery Kruger
}
